+++ CHANGELOG.md
@@ -29,6 +29,9 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Removed `flock` from `::nix::fcntl` on Solaris. ([#2082](https://github.com/nix-rust/nix/pull/2082))
 - Use I/O safety with `copy_file_range`, and expose it on FreeBSD.
   (#[1906](https://github.com/nix-rust/nix/pull/1906))
+- Added `MSG_WAITFORONE` to `MsgFlags` on Android, Fuchsia, Linux, NetBSD,
+  FreeBSD, OpenBSD, and Solaris.
+  ([#2014](https://github.com/nix-rust/nix/pull/2014))
 
 ### Changed